Loop between "ready" and result screens.  Refactoring, including moving stats and adventure functions into their own files to clean up adhop.py.
